Add files via upload
Project Description: Android Phone Forensics and Data Extraction
The primary objective of my project is to unlock an Android phone and extract the maximum amount of data from it for forensic analysis. The device in question belongs to an individual who has been apprehended for allegedly misappropriating funds from the office accounts. As part of the investigation, it is crucial to retrieve and analyze data from the suspect’s phone to determine how the theft was carried out.